requirements.txt
azure-functions
azure-identity
azure-ai-openai
azure-cognitiveservices-speech
requests
python-dotenv
